requirements.txt
# ArtFlow Studio - Python Dependencies

# UI Framework
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# OpenGL bindings
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7

# Vulkan bindings
vulkan>=1.3.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# ABR brush file parsing
construct>=2.10.0

# Database
sqlalchemy>=2.0.0

# HTTP requests (for downloads)
requests>=2.31.0
aiohttp>=3.9.0

# YouTube API
google-api-python-client>=2.100.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# C++ bindings
pybind11>=2.11.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
